+++ Proiecte/Radix_reale/Radix_reale/Source.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int main()
 {	
-	vector<float> test = { 54.34,66.45,98.21,26.33,66.46,66.44 };
+	vector<double> test = { 54.34,66.45,98.21,26.33,66.46,66.44 };
 	RadixSort radix1(test);
 	getchar();
 	return 0;
